raspi_server/requirements.txt
numpy>=1.0,<2.00
Pillow==4.2.1
thriftpy==0.3.9
picamera==1.12
future==0.16.0
Flask>=0.12,<0.13
pytest==3.2.1